add start drag with mirror
This new function startWithMirror may be called directly.
The drag operations starts immediately, following the mouse. It does not need the mouse to be pressed.
The mirrorContainer passed on the factory options is used to render the dragged element visualization.
This is useful if the drag operation originates outside the scope of the drag client and must look like it continues here.
Using a hidden mirrorContainer allows the external host to attach its own visualization.